[WIP] Add performFiles

+++ packages/web/src/apis/localDb.js
@@ -1,9 +1,12 @@
 import * as idb from 'idb-keyval';
 
 import {
-  NOTES, SSLTS, SETTINGS, INFO, PINS, TAGS, UNSAVED_NOTES, DOT_JSON,
+  NOTES, SSLTS, SETTINGS, INFO, PINS, TAGS, UNSAVED_NOTES, DOT_JSON, PUT_FILE,
+  DELETE_FILE,
 } from '../types/const';
-import { isObject, copyFPaths, addFPath, deleteFPath } from '../utils';
+import {
+  isObject, isString, copyFPaths, addFPath, deleteFPath,
+} from '../utils';
 import { cachedFPaths } from '../vars';
 
 // Need cache to work even without IndexedDB.
@@ -30,15 +33,6 @@ const getFile = async (fpath, dangerouslyIgnoreUndefined = false) => {
   return content;
 };
 
-const getFiles = async (fpaths, dangerouslyIgnoreUndefined = false) => {
-  const contents = [];
-  for (const fpath of fpaths) {
-    const content = await getFile(fpath, dangerouslyIgnoreUndefined);
-    contents.push(content);
-  }
-  return { fpaths, contents };
-};
-
 const putFile = async (fpath, content) => {
   if (fpath.endsWith(DOT_JSON)) content = JSON.stringify(content);
 
@@ -58,12 +52,6 @@ const putFile = async (fpath, content) => {
   if (fpath.startsWith(UNSAVED_NOTES)) cachedContents[fpath] = content;
 };
 
-const putFiles = async (fpaths, contents) => {
-  for (let i = 0; i < fpaths.length; i++) {
-    await putFile(fpaths[i], contents[i]);
-  }
-};
-
 const deleteFile = async (fpath) => {
   try {
     await idb.del(fpath);
@@ -80,10 +68,45 @@ const deleteFile = async (fpath) => {
   delete cachedContents[fpath];
 };
 
-const deleteFiles = async (fpaths) => {
-  for (const fpath of fpaths) {
-    await deleteFile(fpath);
+const performFile = async (data) => {
+  const { id, type, path: fpath } = data;
+
+  if (type === PUT_FILE) {
+    const publicUrl = await putFile(fpath, data.content);
+    return { success: true, id, publicUrl };
   }
+
+  if (type === DELETE_FILE) {
+    await deleteFile(fpath);
+    return { success: true, id };
+  }
+
+  throw new Error(`Invalid data.type: ${data.type}`);
+};
+
+const performFiles = async (data) => {
+  const results = [];
+
+  if (Array.isArray(data.values) && [true, false].includes(data.isSequential)) {
+    for (const value of data.values) {
+      const pResults = await performFiles(value);
+      results.push(...pResults);
+      if (data.isSequential && pResults.some(result => !result.success)) break;
+    }
+  } else if (isString(data.id) && isString(data.type) && isString(data.path)) {
+    try {
+      const result = await performFile(data);
+      results.push(result);
+    } catch (error) {
+      results.push({
+        error: error.toString().slice(0, 999), success: false, id: data.id,
+      });
+    }
+  } else {
+    console.log('In localDb.performFiles, invalid data:', data);
+  }
+
+  return results;
 };
 
 const deleteAllFiles = async () => {
@@ -154,8 +177,8 @@ const canUseSync = async () => {
 };
 
 const localDb = {
-  cachedFPaths, getFile, getFiles, putFile, putFiles, deleteFile, deleteFiles,
-  deleteAllFiles, listFiles, exists, getUnsavedNoteFPaths, canUseSync,
+  cachedFPaths, getFile, performFiles, deleteAllFiles, listFiles, exists,
+  getUnsavedNoteFPaths, canUseSync,
 };
 
 export default localDb;
